double to decimal

+++ Tulpep.PayULibrary/Models/Response/Response_Cross/Response_PMTAXADMINISTRATIVEFEE.cs
@@ -2,7 +2,7 @@
 {
     public class Response_PMTAXADMINISTRATIVEFEE
     {
-        public double value { get; set; }
+        public decimal value { get; set; }
         public string currency { get; set; }
     }
 }
